removed unnecessary error checking, merged same-role scripts into one

+++ tools/create-container.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 _sdir=$(dirname "$(readlink -f "$0")")
 _rel="${_sdir##$PWD}"
-set -u
+set -ue
+
+is_on_btrfs(){
+    stat -f --format=%T "$1" | grep -q btrfs
+}
 
 LXC_PATH="/var/lib/lxc"
 container_name="fc"
@@ -9,11 +13,12 @@ container_name="fc"
 show_help(){
     cat <<HELP
 
-    $(basename $0) [options] 
+    $(basename $0) [options] --host debian|arch 
 
     Options:
 
         --name NAME : Container name (default: $container_name)
+        --host HOST : Host distro.
 
 HELP
 }
@@ -33,6 +38,7 @@ help_die(){
 
 # Parse command line arguments
 # ---------------------------
+host=
 # Initialize parameters
 # ---------------------------
 args_backup=("$@")
@@ -48,6 +54,9 @@ while [ $# -gt 0 ]; do
         # --------------------------------------------------------
         --name) shift
             container_name="$1"
+            ;;
+        --host) shift
+            host="${1:-}"
             ;;
         # --------------------------------------------------------
         -*) # Handle unrecognized options
@@ -66,29 +75,30 @@ done; set -- "${args_backup[@]-}"
 # "$@" is in the original state,
 # use ${args[@]} for new positional arguments  
 
-[[ -d "$LXC_PATH/$container_name" ]] && die "Container $container_name already exists."
+[[ -d "$LXC_PATH/$container_name" ]] && die "Container $container_name already exists." || true
+[[ -z ${host:-} ]] && help_die "Host distro must be declared."
 
 [[ "$(whoami)" == "root" ]] || { sudo "$0" "$@"; exit 0; }
 
-is_on_btrfs(){
-    stat -f --format=%T "$1" | grep -q btrfs
-}
-
 is_on_btrfs "$LXC_PATH" && bdev="-B btrfs" || bdev=""
 
-set -x 
-apt-get install lxc debian-keyring debian-archive-keyring
-lxc-create -n $container_name -t debian $bdev -- -r buster # might not work: --packages 
-set +x
-if lxc-start -n $container_name; then 
-    echo "$container_name successfully created."
-    while :; do
-       timeout 3 lxc-stop -n $container_name && break
-       echo "Retrying stopping $container_name"
-       sleep 2
-    done
-else
-    die "Couldn't start $container_name."
-fi
+case $host in
+    debian) 
+        set -x 
+        apt-get install lxc debian-keyring debian-archive-keyring
+        lxc-create -n $container_name -t debian $bdev -- -r buster  
+        set +x
+        ;;
+    arch)
+        set -x 
+        pacman -Sy lxc debootstrap 
+        lxc-create --name=$container_name --template=download -- --dist debian --release buster --arch amd64
+        set +x
+        ;;
+    *)
+        >&2 echo "Currently LXC container creation on $host host is not automated."
+        die "Please manually create a Debian Buster container with name: $container_name"
+        ;;
+esac
 
 echo "Container $container_name has been setup."
